Final Major Code Changes
Now the app is basically ready to present. Other than the adding rating feature which does not seem to work in conjunction with the project, all of the features we have implemented for the vertical prototype are present. Some changes may still be done to the "Init Recipes" class to add or modify recipes for the demo / video.